corrected imperfections

+++ src/answerModule.js
@@ -6,12 +6,16 @@ const collection = database.collection("answers");
 async function addAnswer(answerId, title, question_id) {
     const connection = await db.connect()
     try {
+        const findId = await collection.findOne({answerId})
+        if (findId) {
+            throw new Error("Cet identifiant existe déjà")
+        }
         const data = {answerId, title, question_id}
         
         const result = await collection.insertOne(data)
         return result
     }catch(error) { 
-        console.log(error)
+        throw error
     } finally {
         await connection.close()
     }
@@ -23,7 +27,7 @@ async function listAnswers() {
         const result = await collection.find({}).toArray()
         return result 
     }catch(error) {
-        console.log(error)
+        throw error
     } finally {
         await connection.close()
     }
@@ -32,12 +36,16 @@ async function listAnswers() {
 async function updateAnswer(answerId, title, question_id) {
     const connection = await db.connect()
     try {
+        const findId = await collection.findOne({answerId})
+        if (!findId) {
+            throw new Error("Cet identifiant n'existe pas")
+        }
         const idUp = {answerId}
         const data = {$set : {title : title, question_id : question_id}}
         const result = await collection.updateOne(idUp, data)
         return result
     }catch(error) { 
-        console.log(error)
+        throw error
     } finally {
         await connection.close()
     }
@@ -46,10 +54,14 @@ async function updateAnswer(answerId, title, question_id) {
 async function deleteAnswer(answerId) {
     const connection = await db.connect()
     try {
+        const findId = await collection.findOne({answerId})
+        if (!findId) {
+            throw new Error("Cet identifiant n'existe pas")
+        }
         const result = await collection.deleteOne({answerId : answerId})
         return result
     }catch(error) { 
-        console.log(error)
+        throw error
     } finally { 
         await connection.close()
     }
