+++ src/employeModule.js
@@ -6,12 +6,16 @@ const collection = database.collection("employes");
 async function addEmploye(employeId, name, role) {
     const connection = await db.connect()
     try {
+        const findId = await collection.findOne({employeId})
+        if (findId) {
+            throw new Error("Cet identifiant existe déjà")
+        }
         const data = {employeId, name, role}
         
         const result = await collection.insertOne(data)
         return result
     }catch(error) { 
-        console.log(error)
+        throw error
     } finally {
         await connection.close()
     }
@@ -23,7 +27,7 @@ async function listEmployes() {
         const result = await collection.find({}).toArray()
         return result 
     }catch(error) {
-        console.log(error)
+        throw error
     } finally {
         await connection.close()
     }
@@ -32,12 +36,16 @@ async function listEmployes() {
 async function updateEmploye(employeId, name, role) {
     const connection = await db.connect()
     try {
+        const findId = await collection.findOne({employeId})
+        if (!findId) {
+            throw new Error("Cet identifiant n'existe pas")
+        }
         const idUp = {employeId}
         const data = {$set : {name: name, role : role}}
         const result = await collection.updateOne(idUp, data)
         return result
     }catch(error) { 
-        console.log(error)
+        throw error
     } finally {
         await connection.close()
     }
@@ -46,10 +54,14 @@ async function updateEmploye(employeId, name, role) {
 async function deleteEmploye(employeId) {
     const connection = await db.connect()
     try {
+        const findId = await collection.findOne({employeId})
+        if (!findId) {
+            throw new Error("Cet identifiant n'existe pas")
+        }
         const result = await collection.deleteOne({employeId : employeId})
         return result
     }catch(error) { 
-        console.log(error)
+        throw error
     } finally {
         await connection.close()
     }
