+++ src/questionModule.js
@@ -6,12 +6,16 @@ const collection = database.collection("questions");
 async function addQuestion(questionId, title, type, survey_id) {
     const connection = await db.connect()
     try {
+        const findId = await collection.findOne({questionId})
+        if (findId) {
+            throw new Error("Cet identifiant existe déjà")
+        }
         const data = {questionId, title, type, survey_id}
         
         const result = await collection.insertOne(data)
         return result
     }catch(error) { 
-        console.log(error)
+        throw error
     } finally {
         await connection.close()
     }
@@ -23,7 +27,7 @@ async function listQuestions() {
         const result = await collection.find({}).toArray()
         return result 
     }catch(error) {
-        console.log(error)
+        throw error
     } finally {
         await connection.close()
     }
@@ -32,12 +36,16 @@ async function listQuestions() {
 async function updateQuestion(questionId, title, type, survey_id) {
     const connection = await db.connect()
     try {
+        const findId = await collection.findOne({questionId})
+        if (!findId) {
+            throw new Error("Cet identifiant n'existe pas")
+        }
         const idUp = {questionId}
         const data = {$set : {title: title, type : type, survey_id: survey_id}}
         const result = await collection.updateOne(idUp, data)
         return result
     }catch(error) { 
-        console.log(error)
+        throw error
     } finally {
         await connection.close()
     }
@@ -46,10 +54,14 @@ async function updateQuestion(questionId, title, type, survey_id) {
 async function deleteQuestion(questionId) {
     const connection = await db.connect()
     try {
+        const findId = await collection.findOne({questionId})
+        if (!findId) {
+            throw new Error("Cet identifiant n'existe pas")
+        }
         const result = await collection.deleteOne({questionId : questionId})
         return result
     }catch(error) { 
-        console.log(error)
+        throw error
     } finally {
         await connection.close()
     }
