+++ src/surveyModule.js
@@ -6,12 +6,16 @@ const collection = database.collection("surveys");
 async function addSurvey(surveyId, name, description, createdAt, employe_id) {
     const connection = await db.connect()
     try {
+        const findId = await collection.findOne({surveyId})
+        if (findId) {
+            throw new Error("Cet identifiant existe déjà")
+        }
         const data = {surveyId, name, description, createdAt, employe_id}
         
         const result = await collection.insertOne(data)
         return result
     }catch(error) { 
-        console.log(error)
+        throw error
     } finally {
         await connection.close()
     }
@@ -23,7 +27,7 @@ async function listSurveys() {
         const result = await collection.find({}).toArray()
         return result 
     }catch(error) {
-        console.log(error)
+        throw error
     } finally {
         await connection.close()
     }
@@ -32,12 +36,16 @@ async function listSurveys() {
 async function updateSurvey(surveyId, name, description, createdAt, employe_id) {
     const connection = await db.connect()
     try {
+        const findId = await collection.findOne({surveyId})
+        if (!findId) {
+            throw new Error("Cet identifiant n'existe pas")
+        }
         const idUp = {surveyId}
         const data = {$set : {name: name, description : description, createdAt : createdAt, employe_id: employe_id}}
         const result = await collection.updateOne(idUp, data)
         return result
     }catch(error) { 
-        console.log(error)
+        throw error
     } finally {
         await connection.close()
     }
@@ -46,10 +54,14 @@ async function updateSurvey(surveyId, name, description, createdAt, employe_id) 
 async function deleteSurvey(surveyId) {
     const connection = await db.connect()
     try {
+        const findId = await collection.findOne({surveyId})
+        if (!findId) {
+            throw new Error("Cet identifiant n'existe pas")
+        }
         const result = await collection.deleteOne({surveyId : surveyId})
         return result
     }catch(error) { 
-        console.log(error)
+        throw error
     } finally {
         await connection.close()
     }
